requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
discord.py==1.7.3
et-xmlfile==1.1.0
multidict==5.2.0
numpy==1.21.3
openai==0.27.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
python-dotenv==0.19.1
tqdm==4.62.3
typing-extensions==3.10.0.2
yarl==1.7.0
